Default to disable hero animation
Flutter's HeroControllerScope (which is created by MaterialApp) doesn't like removing pages as opposed to popping them (in TransitionDelegate), for pages underneath the top page. But popping is necessary for state preservation to work (useful only on web though). Calling markForRemove on a page means it's utterly impossible to preserve it's state as a keep-alive.

Actually, the only problem I've seen with enabled hero animations is just a failed assertion: manifest.toHero != newManifest.toHero
is not true